added a helper structs file
