+++ akd/src/errors.rs
@@ -8,7 +8,7 @@
 //! Errors for various data structure operations.
 use core::fmt;
 
-use crate::node_state::NodeLabel;
+use crate::node_label::NodeLabel;
 
 /// Symbolizes a AkdError, thrown by the akd.
 #[cfg_attr(test, derive(PartialEq))]
